refactor(coediting): merge codemirror.ts into utils.ts
