+++ frontend/src/lib/coediting/utils.ts
@@ -1,3 +1,5 @@
+import type { ChangeSet, EditorState } from "@codemirror/state";
+import { sendableUpdates } from "@codemirror/collab";
 import type { CoeditChange } from "./types";
 
 /** Opal-ish hues that read on both themes, one per peer slot. */
@@ -53,4 +55,24 @@ export function diffToChange(
 /** Apply a range change to a string (UTF-16 offsets). */
 export function applyChange(str: string, c: CoeditChange): string {
   return str.slice(0, c.from) + c.insert + str.slice(c.to);
+}
+
+/** Convert a CodeMirror `ChangeSet` to `CoeditChange[]` (old-doc coords).
+ * One entry per changed span — exactly what `iterChanges` yields. */
+export function changeSetToChanges(cs: ChangeSet): CoeditChange[] {
+  const out: CoeditChange[] = [];
+  cs.iterChanges((fromA, toA, _fromB, _toB, inserted) => {
+    out.push({ from: fromA, to: toA, insert: inserted.toString() });
+  });
+  return out;
+}
+
+/** Length of the confirmed (synced) doc = current doc minus the net length of
+ * un-acked local edits. Inbound op `ChangeSet`s must be anchored against this. */
+export function syncedDocLength(state: EditorState): number {
+  let len = state.doc.length;
+  for (const u of sendableUpdates(state)) {
+    len -= u.changes.newLength - u.changes.length;
+  }
+  return len;
 }
